# src/main/resources/swaggerio-spec-0.0.1.yaml
openapi: 3.1.0
info:
  title: Statistics Stat
  description: This is meant to be a site that will allow the user to browse through
    an infintley<br/> expandable compartent of data and use the properties of the
    data to build different<br/> representations of that data.  e.g. Histograms, Pie/Line
    charts etc..
  contact:
    email: craig@craigramage.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Inferred Url
tags:
- name: catalog-api-controller
  description: Catalog Api Controller
paths:
  "/statsstat/catalog/datasets":
    get:
      tags:
      - catalog
      summary: Retrieve all datasets
      description: Returns all of the datasets in the catalog
      operationId: getAllAvailableDatasetsGET
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Dataset"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - catalog_auth:
        - write:catalog
        - read:catalog
components:
  schemas:
    Dataset:
      title: Dataset
      type: object
      properties:
        lastUpdated:
          type: integer
          format: int64
          example: 1672249350
        name:
          type: string
          example: "Example Dataset Name"
        recordCount:
          type: integer
          format: int64
          example: 24586
        description:
          type: string
          example: "This is some example text for a description of a dataset"
        indexableField:
          type: array
          json:
            name: indexableFields
            wrapped: true
          items:
            $ref: '#/components/schemas/DatasetFieldDefinition'
    DatasetFieldDefinition:
      title: DatasetFieldDefinition
      type: object
      properties:
        fieldName:
          type: string
          example: "Example Field Name"
        datatype: 
          type: string
          example: "String"
        
        
  requestBodies:
    Dataset:
      description: Dataset object
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Dataset'
  securitySchemes:
    catalog_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://localhost/oauth/authorize
          scopes:
            write:catalog: modify your catalog datasets
            read:catalog: read your catalog datasets